Clear the App.jsx from useless functions and states and render the design properly
