+++ src/components/simple/RightSide/RightSide.jsx
@@ -1,4 +1,14 @@
-
+import Forecast from "./Forecast";
+import Weather from "./Weather";
+import ValueRow from "../../base/ValueRow";
+import InfoRow from "../../base/InfoRow";
+import {
+  highTempImg,
+  lowTempImg,
+  humadityImg,
+  cloudyImg,
+  windImg,
+} from "../../../imports/rowImages";
 
 export default function RightSide({ children }) {
   return (
@@ -8,7 +18,16 @@ export default function RightSide({ children }) {
         <section className="weather-details-container">
           <p id="weather-details">Weather Details...</p>
         </section>
-        {children}
+        <Forecast forecastHeading={"Thunderstorm with React"}>
+          <ValueRow parameter={"Clouds"} value={"15%"} image={highTempImg} />
+          <ValueRow parameter={"Clouds"} value={"15%"} image={lowTempImg} />
+          <ValueRow parameter={"Clouds"} value={"15%"} image={humadityImg} />
+          <ValueRow parameter={"Clouds"} value={"15%"} image={cloudyImg} />
+          <ValueRow parameter={"Clouds"} value={"15%"} image={windImg} />
+        </Forecast>
+        <Weather>
+          <InfoRow forecastHeading={"Rainy day"} temperature={"16*"} />
+        </Weather>
         <script src="index.js"></script>
       </main>
     </>
